Moved media to assets, removed bg color of body, changed an old scss variable'
