+++ src/assets/covidgif.gif
