+++ src/assets/symptomsImages/symptom_01.png
