+++ src/assets/symptomsImages/symptom_02.png
