+++ src/assets/symptomsImages/symptom_03.png
