+++ src/assets/symptomsImages/symptom_04.png
