+++ src/assets/symptomsImages/symptom_05.png
