+++ src/assets/symptomsImages/symptom_06.png
