+++ src/assets/symptomsImages/symptom_07.png
