+++ src/assets/symptomsImages/symptom_08.png
