+++ src/assets/symptomsImages/symptom_09.png
