+++ src/assets/symptomsRoutes.js
@@ -1,15 +1,15 @@
 // * Covid-19 Images
-import symptom_01 from '../media/symptomsImages/symptom_01.png';
-import symptom_02 from '../media/symptomsImages/symptom_02.png';
-import symptom_03 from '../media/symptomsImages/symptom_03.png';
-import symptom_04 from '../media/symptomsImages/symptom_04.png';
-import symptom_05 from '../media/symptomsImages/symptom_05.png';
-import symptom_06 from '../media/symptomsImages/symptom_06.png';
-import symptom_07 from '../media/symptomsImages/symptom_07.png';
-import symptom_08 from '../media/symptomsImages/symptom_08.png';
-import symptom_09 from '../media/symptomsImages/symptom_09.png';
+import symptom_01 from './symptomsImages/symptom_01.png';
+import symptom_02 from './symptomsImages/symptom_02.png';
+import symptom_03 from './symptomsImages/symptom_03.png';
+import symptom_04 from './symptomsImages/symptom_04.png';
+import symptom_05 from './symptomsImages/symptom_05.png';
+import symptom_06 from './symptomsImages/symptom_06.png';
+import symptom_07 from './symptomsImages/symptom_07.png';
+import symptom_08 from './symptomsImages/symptom_08.png';
+import symptom_09 from './symptomsImages/symptom_09.png';
 
-export const symptomsRoutes = (option) => {
+export const symptomsRoutes = option => {
   switch (option) {
     case 1:
       return symptom_01;
