+++ src/styles/base/base.scss
@@ -15,7 +15,7 @@ html {
   font-family: 'Montserrat', sans-serif !important;
   min-height: 100vh;
   width: 100%;
-  background-color: #4f98ca;
+  // background-color: #4f98ca;
 }
 
 #root {
