+++ src/styles/pages/formPatient.scss
@@ -10,7 +10,7 @@
     text-align: center;
   }
   &__subtitle {
-    border-bottom: 2px solid $secondaryColor;
+    border-bottom: 2px solid $accent;
   }
   &__subtitle {
     text-align: center;
